chore: specify more paths in `.prettierignore` (#10933)
Prettier is a tool with questionable defaults and performance. Not only
is it slow to format files, but it also doesn't respect nested
`.gitignore` or `.prettierignore` files. This means we need to specify
all paths that it should not touch in the top-level .`prettierignore` to
get even close to acceptable performance. Even with these optimisations,
it still takes around 8 seconds to format our workspace.