fix(apple): await all iterations of log export loop (#10996)
A regression was introduced in #10855 that introduced a `Task {}`
closure upon each iteration of the log export loop that was never
awaited. This meant that the log export IPC call returned after the
first chunk was written, which triggered the `defer` block which stopped
the tunnel and caused the next iteration of the task to return `nil`
data, corrupting the log archive.

To fix this, we ensure the ipc calls happen within the loop, passing
control flow logic back to the caller to ultimately determine if further
IPC calls to grab logs should be made.

No changelog since not released yet.