fix(apple): store debug token in separate keychain item (#10994)
When building / testing for debug, it's helpful to have the token stick
around so that we can test autoconnect and toggling the VPN on/off from
system settings.

To facilitate this we need to ensure the release and debug keychain
items do not conflict. There was a previous attempt at handling this,
however, not enough of the keys were differentiated to ensure this
worked correctly.

To store these as separate items, it appears the `account` attribute
should be differentiated as well.

Related: #10986