chore(connlib): capture more data for unroutable packets (#10987)
Certain packets cannot be routed by the Gateway. For those, we emit an
`UnroutablePacket` error. Right now, the information in those is
rendered as "unknown" if we cannot parse the L4 protocol from it. In
order to make the logs more useful, we extend this to capture at least
the IP addresses and the L4 protocol number.