test(connlib): cover more ground with determinism tests (#10974)
In order for the `tunnel_test` test-suite to be useful, we need to
ensure that they are deterministic based off the seed. We have
unit-tests for those but those only check a single `Transition` for a
few hard-coded seeds.

To ensure we don't have some hidden non-determinism somewhere, we
convert that into a proptest and also create a series of transitions the
same way `tunnel_test` does.