fix(relay): allow relay to function on macOS (#11034)
When performing local sanity testing and also doing major refactoring
(especially of the channel modules / wire protocol), it's immensely
helpful to have a miniature Firezone cluster running locally before
deploying to staging and incurring the massive iteration time hit (also
can't deploy to staging if things are broken).

To fix this, we implement the signals used in the relay for `macOS`
which is a POSIX-compliant operating system that should (in theory)
behave the same as Linux.