fix(portal):  support legacy token context types (#11054)
In #10505, we cleaned up tech debt to rename the following token types:

- `gateway_group` to `site`
- `relay_group` to `relay`

The problem with this is that tokens generated prior to this migration,
when decoded, yield the legacy token type strings.

To fix this, we add fallback clauses to support these legacy token
types, reverting the `STATIC_SEEDS`-generated tokens in CI to the
pre-migration values as a safety check to ensure old tokens still work.

Fixes #11046