fix(gui-client): don't panic if `$XDG_RUNTIME_DIR` is not set (#11075)
It appears that on some systems, the `$XDG_RUNTIME_DIR` is not set. In
that case, the GUI cannot function because we don't have a well-known
place to put our IPC sockets.

Instead of panicking, we now return an error from the `ipc_path`
function.

Whilst investigating the behaviour around these errors, I noticed that
we are currently always assuming that the GUI is already running if
creating the GUI IPC server creation fails. That is not correct. We fix
this by forwarding all errors from the GUI IPC server and return
`AlreadyRunning` only in the case when we successfully handshaked with
the existing session.