fix(android): only select apps that are browsable (#11077)
When a user has a browser installed that doesn't support custom tabs, we
fallback to just opening the Auth URL via an `ACTION_VIEW` intent. This
can select non-browser apps if they are registered to handle HTTP links.
To fix this, we need to restrict the intent to apps in the "browsable"
category.

While testing this, I noticed that we will still crash if there is no
browser available. Hence, I've also added a check for that and removed
some dead code in the module.

Source: https://stackoverflow.com/a/67485825