refactor(portal): move relay tokens to separate table (#11101)
Moves the relay token type to a dedicated table, preserving all existing
tokens in the process.

All fields unused / unneeded by relay tokens are removed for simplicity
/ clarity.

Additionally a `create_relay_token/0` helper is added to Domain.Auth to
make it easy to create these for deployments if needed.

Fixes #11094