fix(phoenix-channel): try all addresses in parallel (#11115)
Some systems may report IPv6 connectivity without actually being able to
route IPv6 packets. Normally, such situations result in ICMP messages
which then fail the `connect` operation with an IO error. If that
doesn't happen, the `connect` operation may hang and run into our 5s
timeout for making a connection without ever trying the other addresses.

To fix this, we implement happy eyeballs for the WebSocket connection
which attempts all addresses in parallel, picking the one which resolves
first and dropping the other sockets.

This ensures we are able to create a socket connection even if some
paths are entirely non-responsive.