fix(portal): use worker states to allow subsequent runs (#11138)
Since the introduction of the
[lifeline](https://hexdocs.pm/oban/2.9.1/lifeline.html) plugin, we have
a hard cap of 2 hours before jobs stuck in the `executing` state are
"rescued" and available for running again.

This means that we should effectively allow a uniqueness period of
`:infinity` for jobs in the executable states. Otherwise, a job could be
started again after the `period` even if it is already executing.