chore(portal): Add elixir code coverage config (#11128)
Why:

* With the recent large refactor we're now in a place of having many
tests that are either irrelevant or are not passing. This will require
rewriting many tests in a lot of cases. In this commit I've moved all of
the test files that have some type of failure to `.old` I've also made
sure that the code coverage tooling is configured for our elixir app.
This should allow us to have a quick gauge of where we're lacking as we
rewrite tests for all of the elixir apps.

Related: #11056