test(portal): allow passing in auth_provider assoc
Allows passing in the base auth_provider association when creating a
child auth provider. This is useful when you need a base auth_provider
to use for an association (like portal_session) and want a child
provider associated with it.

Without this, the base auth_provider association will not be loaded on
the record returned by `email_otp_provider()` function call for example.