fix(connlib): only emit resource list on changes (#11129)
To reduce the churn in the Client apps, we now accumulate all updates to
the resource list and only emit an event if it has actually changed
since the last one we emitted. To do this reliable, we introduce the
`TrackedState` component.

A `TrackedState` combines a current value with a `PendingUpdate`. By
default, both are empty. Consecutive calls to `update` create a kind of
"Transaction" or "Session" tracked by the `PendingUpdate` struct. At any
point, the `PendingUpdate` can be consumed which compares the hash-code
of the state prior to the first update with the final, desired value. An
update is then only emitted if anything has actually changed.

This logic ensures that even a series of updates, like multiple messages
from the portal in a row do not result in a Client side update of the
resource list or TUN config if nothing has actually changed.

Related: #11120