refactor(portal): move browser tokens to portal_sessions (#11127)
Moves the `browser` token type to a new table, `portal_sessions` and
updates all references.

These are added to the change_log for now to ensure if we want them in
there in the future, we'll have them retroactively.

Updates associated tests that have already been created, but does not
create new controller test files just yet.

Related: #11056
Fixes: #11100