refactor(windows-event-log): remove conditional compilation (#11160)
The current implementation of our windows event log integration only
compiles on Windows. Whilst it can never work on Windows, we can
structure the code such that it can at least be integrated on other
platforms as a no-op module. This simplifies downstream code because it
doesn't have to wrestle with `cfg`s.

Additionally, we remove the built-in event filtering support. Tracing
layers are meant to be composed and filtering is something that can
easily be done on top.

Resolves: #11135