fix(portal): allocate tun addresses in O(1) time (#11154)
When inserting new Clients and Gateways, we need a unique IPv4 and IPv6
address within our designated ranges to reserve to use for the tunnel
interface IPs.

This is unfortunately not a trivial problem to solve. It requires the
use of a shared `network_addresses` table which is used to ensure unique
addresses across both entity types with a foreign key from `(account_id,
address)` on network_addresses to `(account_id, ipv4)` on
clients/gateways.

To find a free address, we used an algorithm that involved:

1. Opening a transaction
2. Run a query to scan for available addresses in the
`network_addresses` table by picking a random offset into the range and
then scanning forwards *and* backwards for the first unused IP
3. Grab it 
4. Perform the insert for both network_address and client/gateway

This meant that as occupancy grows, the amount of time we hold open the
transaction (and risk for contention) grows as well.

To reduce the average-case runtime of this allocation from O(n) to O(1)
we implement the following algorithm:

- Upon insert, find the max used address, increment and save counter
then attempt to insert address
- If success, return address
- If collision, increment and try again
- If num_iterations > num_addresses_in_cidr, we've exhausted the pool -
raise an exception

The tradeoff here is that addresses are handed out with a sequential
pattern instead of the prior random pattern, but random address
allocation was never a requirement for this particular function and
sequential allocation should be fine.

Fixes #11153 
Fixes #10151