fix(apple): accumulate changes to `NetworkSettings` (#11191)
Right now, the order in which the `ResourcesUpdated` and
`TunInterfaceUpdated` events are processed by the Network Extensions
matters. That is because the `NetworkSettings` only get initialized on
the first `TunInterfaceUpdated` event and hence the data from prior
`ResourcesUpdated` events is lost until it is emitted again.

This causes user-visible issues because we re-apply the
`NetworkSettings` on changes to the resource list to flush the DNS
cache. When a user first accesses a resource, we connect to the
corresponding site and therefore change the site status to "Online".
Because prior updates of the resource list have been discarded, this is
the first event that informs the `NetworkSettings` of the list of DNS
resources and it gets re-applied to clear the cache, causing browsers
like Chrome to reset all of their sockets.

To fix this, we rework the `NetworkSettings` class to a simple struct
that just holds and accumulates data. It internally performs diffs on
every update and returns an instance of `NEPacketTunnelNetworkSettings`
if anything has changed.

This fixes this issue because `NEPacketTunnelNetworkSettings` can only
be built if we have IP addresses for the TUN device and the list of
resources gets buffered internally as the current list. Hence,
subsequent updates to the resource list won't create a new
`NEPacketTunnelNetworkSettings` and thus won't disrupt the user in
loading the resource.

To verify this further, we implement unit-tests and run them in CI.

Fixes: #11185

---------

Signed-off-by: Thomas Eizinger <thomas@eizinger.io>