chore: Add Rust lint to pre-commit (#11198)
Start linting custom Rust rules with pre-commit.

In Rust, we favour using `tracing::` prefix instead of importing the
macros.
Add a custom rule and fix the errors as part of this changeset.

Also, as part of development of the rule, and to make it more useful for
other devs, improve pre-commit configuration to:
1. Only check files tracked by git with prettier (avoids failing on
local markdown notes etc)
2. Use shellcheck directly (it's already specified by .tool-versions)
instead of via Docker container - matters on macOS, where Docker is not
as lightweight.

---------

Signed-off-by: Mariusz Klochowicz <mariusz@klochowicz.com>
Co-authored-by: Thomas Eizinger <thomas@eizinger.io>