feat(gateway): report event-loop errors to Sentry (#11209)
The Gateway doesn't print errors from `try_main` on the ERROR level
because it primarily spammed Sentry with what are effectively
user-errors, for example starting the Gateway without root.

When we make it all the way to the event-loop, it would be useful to
know if the Gateway fails. To do that, we add a custom error context and
if present, we log the error message on the ERROR level, which ends up
getting forwarded to Sentry.