fix(portal): don't double-prepend nonce
In #11166 we made a subtle yet breaking change to prepend the
`nonce <> "."` to the signed output in `encode_fragment!`.

The encoded fragment should be the signed portion of the token without
the nonce prepended. Clients receiving tokens via insecure channels
(i.e. GUI clients) receive this signed fragment and then prepend their
nonce to it, forming the full token.

A test is also added to ensure this regression doesn't resurface.