fix(portal): ensure client has preloaded addresses (#11205)
In #11154 we moved the ipv4/6 address to another table. What we failed
to do however was ensure these fields are preloaded before rendering the
client down to the gateway. This would cause an issue if the client is
updated in the client channel from a DB Change hook, since we only
receive the immediate record in those handlers.

This wouldn't occur if the client was never updated, because we preload
the associations when hydrating the initial data.

Fixes #11201