fix(portal): migrate existing google/okta groups
Unfortunately the migration added in #10505 only migrated existing Entra
groups, meaning that when setting up Google or Okta sync, new groups
were created instead of connecting/upserting the existing ones.

This migration fixes that issue for staging and also prod by stripping
the prefixes and setting the `entity_type` based on the current value of
the `idp_id`.

On staging, this will cause a conflict since we have already set up the
new adapters. To handle conflicts gracefully, we treat the existing,
prefixed group entities as the "source of truth" and if a conflict
arises, we delete the newer one which won't have been to linked to any
policies.

Fixes #11207