fix(portal): use consistent callback url for verification
When verifying Entra auth and sync configurations, we were using a
different callback URI, `/verification` since Entra handles things using
a special OIDC flow dubbed "admin consent".

This is actually a flavor of OIDC under the covers, so rather than have
a special handler for Entra, we now handle it in the
`/auth/oidc/callback` handler like other verifications.

Furthermore, we simplify the key that the verification session data is
stored under to be a single `verification` key such that only one
verification can be in flight at a time.

Fixes #11189