ci(rust): only run tests on 1 version per OS (#11229)
Our Rust code is primarily cross-platform and for the platform-specific
bits, we don't have many tests. The biggest part here is our UDP socket
which is already very well-tested in the upstream `quinn` repository. As
such, running all of our Rust tests on several versions of each
operating system is just a waste in CI resources, especially because we
only have 5 macos runners across the entire organization.