ci: cancel entire run if any workflow fails in merge-queue (#11232)
Right now, all workflows defined in our `ci.yml` workflow are
independent of each other. That means, they will all independently run
to completion, regardless of whether any others have already failed.
This unnecessarily hogs up resources, especially in the merge-queue.

At least in the merge-queue, GitHub would recognize this if we would
define all our jobs as "Required". GitHub's UI for that is super
cumbersome though which is why we only have our aggregate
`required-check` job.

In order to be more resource-efficient, we add a `monitor` job for each
embedded workflow. The monitor jobs runs as a dependency of its workflow
and cancels the entire run if it failed, thus freeing up resources for
other CI runs, effectively de-queueing the PR early from the
merge-queue.