fix(gateway): retry binding DNS servers on TUN interface (#11208)
We currently have a bug where the Gateway may restart if the WebSocket
gets cut and we need to rebind the DNS servers on the TUN interface. The
sockets may not have been fully dropped or the interface may not be
fully available yet. Additionally, the last release introduced a
regression where the Gateway would not start up if IPv6 was completely
disabled on the machine.

To fix this, we now return a `TunIpStack` enum from
`TunDeviceManager::set_ips` which allows us to further inform, which
operations we want to make and only try to bind DNS servers on the IPs
that worked. We retry this 3 times in case the previous sockets haven't
been fully cleared yet.