fix(portal): don't redirect authenticated for as=client (#11206)
When determining whether to redirect a sign-in flow based on whether
we're already authenticated, we weren't taking into account whether the
flow was for a client. If we already had an active session to the portal
opened in the same browser, we would land straight into the portal when
authenticating as a client.

Fixes #11203