feat(portal): expose all group fields (#11238)
When viewing groups in the UI, we weren't displaying any indication
whether the group was an org_unit or not. Since 99% of all groups are
going to be `group`s and not org_units, we add an indication in both the
index view and the modal only when this is set to `org_unit`.

Additionally, all fields except the internal `type` field have been
added to the REST API, and we clean up the check that determines what's
a synced group based off `idp_id`. This field will be set for migrated
groups that are not yet re-attached to a directory, so we don't want to
allow modifying these either.

<img width="1056" height="684" alt="Screenshot 2025-12-17 at 8 17 05 AM"
src="https://github.com/user-attachments/assets/687f0256-c4c0-4435-a751-98f54d86e855"
/>
<img width="709" height="749" alt="Screenshot 2025-12-17 at 8 17 15 AM"
src="https://github.com/user-attachments/assets/15d1efa4-f00b-4fd5-8c0b-dcf590cdd160"
/>


Fixes #11220