ci: only run performance benchmarks on main (#11245)
Building release binaries for all data plane components on every CI run
is quite costly. We are already building them when merge to `main` so
running the performance tests only there gives us basically the same
amount of data in bencher, i.e. an entry for every commit that gets
pushed to `main`. We don't learn ahead of time when a PR is about to
introduce a regression but at least right now, the results from perf
runs on GitHub actions are so variable that we cannot make use of those
alerts anyway.

This should free up a significant amount of CI resources and remove
about a dozen jobs from each merge-queue run.