ci: add `!cancelled()` condition to monitor jobs (#11251)
It appears that without the additional `!cancelled()` condition, the
"monitor" jobs get cancelled because the workflow they are depending on
failed. Running them _on failure_ is the entire point though so we now
add `!cancelled()` there.

I tested this in this PR by temporarily removing the `merge_group`
condition and it now works as expected.