fix(portal): apply debouncing to all text inputs (#11248)
Adding `phx-debounce` to a `form` does not debounce its child `input`
changes. This means we were issuing updates on each keypress / click in
all of our live table filters.

We fix this by adding `phx-debounce="300"` to the inputs in the
live_table implementation. We also add it to the group filter input in
the policies forms.

Lastly, we add it to various other forms where we aren't neecessarily
hitting the DB but are performing validations in order to prevent
spurious roundtrips.

Fixes #11052