ci: retry docker remote operations 5 times (#11254)
Most cases of CI flakiness happen due to Docker remote operation
timeouts during either the login step or the remote image fetch.

By default, a single failure here means that particular job fails,
taking the entire workflow run down with it.

To alleviate this, we allow retries up to a maximum of 5 attempts for
the `ghcr-docker-login` action and remote operations in the
`setup-docker` action.

There are other places where docker is used where these won't be
applied, but hopefully this serves as a starting point and we can add
more retries on an as-needed basis.