feat(connlib): don't concurrently create flows for new sites (#11266)
When connlib receives a packet for a resource that it isn't connected to
yet, we send a `create_flow` message to the portal. Right now, those are
only de-duplicated and debounced on a per-resources basis, i.e. if we
have sent a `create_flow` message for a given resource in the last two
seconds already, we suppress it in order to not overload the portal.

When we aren't connected to any resources yet and the user signs into
Firezone, we very likely need to authorize connections for multiple
resources in parallel. When we aren't connected to any sites yet, this
creates a race condition.

When sending a `create_flow` message, we also include the Gateways we
are already connected to. The portal then prefers those instead of
connecting us to a new Gateway. But when sending `create_flow` messages
for multiple resources in the same site in parallel, all of them will
report no connected Gateways which may lead to us connecting to several
Gateways.

To fix this, we hold back connection intents for resources in the same
site if we aren't connected to the site yet and have a "pending flow"
locally. Packets for these resources are still buffered and will be
flushed once the flow for this resource is authorized.

To better test this logic, we extract the `PendingFlows` component which
represents the state machine around these connection intents.