fix(portal): use gin indexes for fulltext_search
In nearly all of the places in the portal we expose a text input to
filter by name or address in a live table, we were using naive `ilike`
from Ecto.Query.

Without an index, this does a sequential scan on the tables involved.
Depending on the size of the table (e.g. clients) this could be a
substantial load on the DB.

To alleviate this, we:

- Simplify our fulltext_search function to only use trigram queries
  using the `unaccent` extension. This allows for partial matches on
  words like `cafe` when searching for `afe` for example.
- Add GIN indexes to all colums exposed for this

Fixes #11260