fix(connlib): replace `expect` with WARN log (#11275)
Panicking in connlib is a bad idea, even for "this should never happen"
cases because it takes down the entire data plane and is really
disruptive for the user. Meanwhile, having the UI not update, which
sites we are connected to is at best annoying.

By logging a warning instead, we are on the safer side here and can fix
any bug that comes up in a follow-up release.