ci: split `setup-rust` action (#11268)
Currently, `setup-rust` does several things, not all of which are needed
for all workflows. Additionally, it always sets up caching via sccache.

Sccache is very useful when you have a diverse number of jobs, all of
which compile slightly different subsets of our packages. Its shared
nature provides a space-efficient trade-off that still allows you to
cache things in such a scenario. The downside is that it has to perform
a network request for each cached compilation artifact that it wants to
fetch.

In some scenarios, having the exact crates that need to be built already
on disk is significantly faster.

Our data-plane components are needed in order to run our integration
tests. For PRs that don't change Rust code, those need to be built with
the Rust code that is on `main`. To make these builds faster, we split
the `setup-rust` action up such that we can configure "regular" GitHub
actions-based caching of the `target` directory for these jobs. This
isn't very space-efficient but it greatly speeds up the workflows if we
create a dedicated cache per job.

With this configuration, the data plane components for the integration
tests build in just under 2 minutes.