fix(connlib): don't allow more than 1 connection per site (#11267)
Within connlib, we assume in several places that we are only connected
to a single Gateway per site. For example, in order to maintain Gateway
affinity, we send the list of currently connected Gateways to the portal
with each `create_flow` message. If we somehow end up being connected to
two Gateways in the same site, this may result in the portal giving us a
Gateway for a resource that we previously routed to a different Gateway.

This is an issue because Firezone needs to be as transparent as possible
to all network traffic. Ultimately, the Gateway's public IP will be
observed as the source of the traffic by the resource. Multi-path
protocols are technically possible and QUIC may support this in the
future. In today's Internet though, most network connections cannot be
roamed to a different endpoint. As such, it is important that we don't
move connections between Gateways unless absolutely necessary (like if
the admin moves a resource to a different site).