ci: shorten cooldowns to 2, 7, 14 days (#11331)
Unfortunately the `28` day cooldown means any package with a release
cadence shorter than 28 days will _never_ be updated.

This PR fixes that by specifying a cascading cooldown period based on
the patch, minor, and major:

- For patch updates, use a 2-day cooldown period
- For minor updates, use a 7-day cooldown period
- For major updates, use a 14-day cooldown period

Security updates are already configured to open ASAP as soon as
Dependabot knows about them. I'm not sure cooldowns are the lever we
want to pull here. We update the grouping for each ecosystem to ensure
we only see _one_ PR to bump security updates to avoid PR spam.

Related: #11294