fix(portal): do not redirect authed legacy provider callbacks (#11401)
Legacy OIDC providers were mistakenly being caught in the
`RedirectIfAuthenticated` plug which took an admin attempting to sign
into a client straight to the admin portal.

To fix this, we move this route out to a `:public` pipeline with no
plugs in the way and beef up the tests to prevent regressions like this
going forward.

Fixes #11203