fix(portal): used signed cookies (#11404)
Encrypted cookies add a decent amount of size overhead and can push over
the limit if we're not careful. Since we are not storing any sensitive
data in these, we can reduce the size of all cookies by signing them
only.

This PR updates our cookie system to:

- Use our app-wide signing salt for all cookies
- Ensure all cookies are signed
- Ensure all cookies are `secure: true` even for dev (`localhost`
bypasses this on modern browsers, so this shouldn't affect developing)
- Renames some of the cookies so as not to clobber with any existing
ones and result in a cookie decode error
- Adds a new struct pattern for each cookie we want to store. Each
struct is responsible for converting itself into binary and term forms
for serialization / compaction.
- Reduces the number of recent accounts we track to 10 to avoid the HTTP
431

Fixes #11400